requirements.txt
langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
chromadb
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime
langchain-groq